feat:file接口修改，环境变量修改

+++ src/routes/files/file_chunkFn.ts
@@ -3,11 +3,13 @@ import path from 'path'
 import fs from 'fs'
 import lodash from 'lodash'
 import { getVideoDurationInSeconds } from 'get-video-duration'
-import { createHash, streamToBuffer } from './utils'
-import { createRequestReturn, getNetworkIp, getProjectPath } from '../../utils'
+import { createHash, getMediaType, streamToBuffer } from './utils'
+import { getLocalIp, getProjectPath } from '../../utils'
 import { Files_chunk } from './file_chunk'
 
 const { isNil } = lodash
+const publicUrl =
+  process.env.NODE_ENV.trim() === 'dev' ? getLocalIp() : process.env.PUBLIC_URL
 
 const basePath = path.join(getProjectPath(), 'public')
 let temp: Files_chunk[] = []
@@ -40,34 +42,37 @@ export async function uploadFileChunk(req: FastifyRequest): Promise<void> {
 export async function mergeFileChunk(data: {
   uuid: string
   fileType: string
-  mediaType: string
+  // mediaType: string
 }): Promise<any> {
   const all: Buffer[] = temp
-    .filter((obj) => (obj.uuid = data.uuid)) // 将参数对应uuid的文件块过滤出来
+    .filter((obj) => obj.uuid === data.uuid) // 将参数对应uuid的文件块过滤出来
     .sort((a, b) => a.currentSlicesNum - b.currentSlicesNum) // 将文件块排好序
     .map((obj) => obj.fileSlices as unknown as Buffer) // 提取出文件块的buffer数据
+  temp = temp.filter((obj) => obj.uuid !== data.uuid) // 删除该文件所有的块
   const allBuffer = Buffer.concat(all) // 拼接buffer数据
   const fileHash = await createHash(allBuffer) // 计算hash值
-
-  temp = temp.filter((obj) => obj.uuid !== data.uuid) // 删除该文件所有的块
+  const mediaType = await getMediaType(allBuffer) // 计算文件类型
 
   const fileName = `${fileHash}.${data.fileType}` // 文件名
-  const filePath = path.join(basePath, data.mediaType, fileName) // 文件绝对路径
-  const fileUrl = `${getNetworkIp()}:${process.env.PORT}/public/${
-    data.mediaType
-  }/${fileName}` // 文件网络路径
-  fs.writeFileSync(filePath, allBuffer)
+
+  const filePath = path.join(basePath, mediaType, fileName) // 文件绝对路径
+  const fileUrl = `http://${publicUrl}:${process.env.PORT}/public/${mediaType}/${fileName}` // 文件网络路径
+  const fileExist = fs.existsSync(filePath)
+  if (!fileExist) {
+    fs.writeFileSync(filePath, allBuffer)
+  }
   const fileSize = fs.statSync(filePath).size
   let result: any = {
     name: fileName,
     url: fileUrl,
-    size: fileSize
-    // mediaType: data.mediaType,
-    // fileType: data.fileType
+    size: fileSize,
+    mediaType,
+    fileType: data.fileType
   }
-  if (data.mediaType === 'videos') {
-    const duration = await getVideoDurationInSeconds(fileUrl) // 如果是视频，计算视频的时间长度
+  if (mediaType === 'videos') {
+    const duration = await getVideoDurationInSeconds(filePath) // 如果是视频，计算视频的时间长度
     result = { ...result, duration }
   }
-  return createRequestReturn(200, result, '')
+
+  return result
 }
